# ALIN_Neustart_Core/25_API_Backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ALIN Local API",
    "description": "Lokales API-Backend für ALIN Legal AI – keine Cloud-Anbindung",
    "version": "0.4.0"
  },
  "paths": {
...
    "/terminologie/{begriff}": {
      "get": {
        "summary": "Terminologie Eintrag",
        "operationId": "terminologie_eintrag_terminologie__begriff__get",
        "parameters": [
          {
            "name": "begriff",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Begriff"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}